pyproject.toml
[project]
name = "lumina-card-exporter"
version = "0.1.0"
description = ""
authors = [
    {name = "Yannou27",email = "yann.tireau27@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=12.1.1,<13.0.0)",
    "fpdf2 (>=2.8.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
